requirements.txt
blinker==1.8.2
click==8.1.7
Deprecated==1.2.14
Flask==3.0.3
Flask-Limiter==3.8.0
gunicorn==22.0.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.1
Pygments==2.18.0
rich==13.7.1
typing_extensions==4.12.2
Werkzeug==3.0.3
wrapt==1.16.0
python-dotenv
